paraphraser-agent/pyproject.toml
[project]
name = "paraphraser-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "grpcio<1.72.1",
    "openai>=1.84.0",
    "openai-agents>=0.0.17",
    "python-dotenv>=1.1.0",
]
